Change Style | Create Job View / Add Functionality
Prior to this commit, there was little functionality to the website. Now, if you go to the charity "Create Job" link, you will find a nicely styled form that will add your information to the database. Furthermore, a user can go to the programmer "Available Jobs" link and see all of the posted charity jobs in a decently styled table.