Create and Apply For Jobs Complete
The authentication system has now been linked with the actual function of the website. A user can now register as a charity or programmer, log in, create a job as a charity, view all created jobs as a  programmer and apply and the charity will see the information about that programmer in their view.