Route Updates | Add Comments
All of the current files have been commented to make them more easily understood. The routes have now been protected by middleware that will check if your are logged in and also check if you are a programmer going to the programmer route and not a charity and vice versa.