Add changes regarding signup and job application
During the registration process, a different set of form data will be shown or hidden depending on what organization (charity / programmer) the user picked. Furthermore, a new screen will be shown when a programmer applies for a job. The screen will show the Charity's Name, Description (all filled out in the registration) , and the purpose of the job. The programmer will respond by typing a "Why I should get this job" paragraph or so in the text area provided and click submit. The charity willl see a list of programmers and their description now in their view.